requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.7.0rc2
dj-rest-auth==2.1.3
Django==3.1.5
django-address==0.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.2
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.11.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
jmespath==0.10.0
msgpack==1.0.2
oauthlib==3.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
slack-sdk==3.3.1
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
zope.interface==5.2.0
